update primitive types to be nullable
This allows us to validate whether the user actually provided a value or
not. This is especially important for booleans since the default value is
false and unless the boolean is nullable there is no way to no whether or
not the user specified false or no value was specified.